backend/requirements.txt
anyio==3.7.1
asgiref==3.9.1
cachetools==4.2.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.68.1
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
greenlet==3.2.4
h11==0.12.0
httpcore==0.13.7
httplib2==0.22.0
httpx==0.19.0
idna==3.10
oauthlib==3.3.1
passlib==1.7.4
psycopg2-binary==2.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.22
PyJWT==2.1.0
pyparsing==3.2.3
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.14.2
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.15.0
alembic==1.14.1

# Testing dependencies
pytest>=6.2.4
pytest-cov>=2.12.1
pytest-asyncio>=0.15.1
